# packages/backend/src/generated/swagger.json
{
    "components": {
...
        "schemas": {
            "ApiErrorPayload": {
                "properties": {
                    "error": {
                        "properties": {
                            "data": {
                                "description": "Optional data containing details of the error"
                            },
                            "message": {
                                "type": "string",
                                "description": "A friendly message summarising the error"
                            },
                            "name": {
                                "type": "string",
                                "description": "Unique name for the type of error"
                            },
                            "statusCode": {
                                "type": "number",
                                "format": "integer",
                                "description": "HTTP status code"
                            }
                        },
                        "required": ["name", "statusCode"],
                        "type": "object"
                    },
                    "status": {
                        "type": "string",
                        "enum": ["error"],
                        "nullable": false
                    }
                },
                "required": ["error", "status"],
                "type": "object",
                "description": "The Error object is returned from the api any time there is an error.\nThe message contains"
            },
...
            "Pick_CreateDbtCloudIntegration.metricsJobId_": {
                "properties": {
                    "metricsJobId": {
                        "type": "string",
                        "description": "Job id for a dbt cloud job containing a compiled dbt project with available dbt metrics"
                    }
                },
                "required": ["metricsJobId"],
                "type": "object",
                "description": "From T, pick a set of properties whose keys are in the union K"
            },
            "DbtCloudIntegration": {
                "$ref": "#/components/schemas/Pick_CreateDbtCloudIntegration.metricsJobId_",
                "description": "Configuration for a Lightdash integration with dbt Cloud"
            },
            "ApiDbtCloudIntegrationSettings": {
                "properties": {
                    "results": {
                        "$ref": "#/components/schemas/DbtCloudIntegration"
                    },
                    "status": {
                        "type": "string",
                        "enum": ["ok"],
                        "nullable": false
                    }
                },
                "required": ["status"],
                "type": "object"
            },
            "ApiDbtCloudSettingsDeleteSuccess": {
                "properties": {
                    "results": {},
                    "status": {
                        "type": "string",
                        "enum": ["ok"],
                        "nullable": false
                    }
                },
                "required": ["status"],
                "type": "object"
            },
...
            "Group": {
                "properties": {
                    "organizationUuid": {
                        "type": "string",
                        "description": "The UUID of the organization that the group belongs to"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "The time that the group was created"
                    },
                    "name": {
                        "type": "string",
                        "description": "A friendly name for the group"
                    },
                    "uuid": {
                        "type": "string",
                        "description": "The group's UUID"
                    }
                },
                "required": ["organizationUuid", "createdAt", "name", "uuid"],
                "type": "object"
            },
            "ApiGroupResponse": {
                "properties": {
                    "results": {
                        "$ref": "#/components/schemas/Group"
                    },
                    "status": {
                        "type": "string",
                        "enum": ["ok"],
                        "nullable": false
                    }
                },
                "required": ["results", "status"],
                "type": "object"
            },
            "ApiSuccessEmpty": {
                "properties": {
                    "results": {},
                    "status": {
                        "type": "string",
                        "enum": ["ok"],
                        "nullable": false
                    }
                },
                "required": ["status"],
                "type": "object"
            },
            "GroupMember": {
                "properties": {
                    "lastName": {
                        "type": "string",
                        "description": "The user's last name"
                    },
                    "firstName": {
                        "type": "string",
                        "description": "The user's first name"
                    },
                    "email": {
                        "type": "string",
                        "description": "Primary email address for the user"
                    },
                    "userUuid": {
                        "type": "string",
                        "description": "Unique id for the user",
                        "format": "uuid"
                    }
                },
                "required": ["lastName", "firstName", "email", "userUuid"],
                "type": "object",
                "description": "A summary for a Lightdash user within a group"
            },
            "ApiGroupMembersResponse": {
                "properties": {
                    "results": {
                        "items": {
                            "$ref": "#/components/schemas/GroupMember"
                        },
                        "type": "array"
                    },
                    "status": {
                        "type": "string",
                        "enum": ["ok"],
                        "nullable": false
                    }
                },
                "required": ["results", "status"],
                "type": "object"
            },
            "Pick_Group.name_": {
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "A friendly name for the group"
                    }
                },
                "required": ["name"],
                "type": "object",
                "description": "From T, pick a set of properties whose keys are in the union K"
            },
            "UpdateGroup": {
                "$ref": "#/components/schemas/Pick_Group.name_"
            },
...
        },
        "securitySchemes": {
            "session_cookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "connect.sid"
            },
            "api_key": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization",
                "description": "Value should be 'ApiKey <your key>'"
            }
        }
    },
    "info": {
        "title": "Lightdash API",
        "version": "0.714.0",
        "description": "Open API documentation for all public Lightdash API endpoints",
        "license": {
            "name": "MIT"
        },
        "contact": {
            "name": "Lightdash Support",
            "email": "support@lightdash.com",
            "url": "https://docs.lightdash.com/help-and-contact/contact/contact_info/"
        }
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/projects/{projectUuid}/integrations/dbt-cloud/settings": {
            "get": {
                "operationId": "getDbtCloudIntegrationSettings",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiDbtCloudIntegrationSettings"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Get the current dbt Cloud integration settings for a project",
                "tags": ["Integrations"],
                "security": [],
                "parameters": [
                    {
                        "description": "the uuid of the project",
                        "in": "path",
                        "name": "projectUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "post": {
                "operationId": "updateDbtCloudIntegrationSettings",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiDbtCloudIntegrationSettings"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Update the dbt Cloud integration settings for a project",
                "tags": ["Integrations"],
                "security": [],
                "parameters": [
                    {
                        "description": "the uuid of the project",
                        "in": "path",
                        "name": "projectUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "delete": {
                "operationId": "deleteDbtCloudIntegrationSettings",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiDbtCloudSettingsDeleteSuccess"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Remove the dbt Cloud integration settings for a project",
                "tags": ["Integrations"],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "projectUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
        "/api/v1/groups/{groupUuid}": {
            "get": {
                "operationId": "getGroup",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiGroupResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Get group details",
                "tags": ["User Groups"],
                "security": [],
                "parameters": [
                    {
                        "description": "unique id of the group",
                        "in": "path",
                        "name": "groupUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "delete": {
                "operationId": "deleteGroup",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessEmpty"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Delete a group",
                "tags": ["User Groups"],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "groupUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "patch": {
                "operationId": "updateGroup",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiGroupResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Update a group",
                "tags": ["User Groups"],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "groupUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateGroup"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/groups/{groupUuid}/members/{userUuid}": {
            "put": {
                "operationId": "addUserToGroup",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessEmpty"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Add a Lightdash user to a group",
                "tags": ["User Groups"],
                "security": [],
                "parameters": [
                    {
                        "description": "the UUID for the group to add the user to",
                        "in": "path",
                        "name": "groupUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "the UUID for the user to add to the group",
                        "in": "path",
                        "name": "userUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "delete": {
                "operationId": "removeUserFromGroup",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessEmpty"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Remove a user from a group",
                "tags": ["User Groups"],
                "security": [],
                "parameters": [
                    {
                        "description": "the UUID for the group to remove the user from",
                        "in": "path",
                        "name": "groupUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "the UUID for the user to remove from the group",
                        "in": "path",
                        "name": "userUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/api/v1/groups/{groupUuid}/members": {
            "get": {
                "operationId": "getGroupMembers",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiGroupMembersResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "View members of a group",
                "tags": ["User Groups"],
                "security": [],
                "parameters": [
                    {
                        "description": "the UUID for the group to view the members of",
                        "in": "path",
                        "name": "groupUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ],
...
}